pyproject.toml
[tool.poetry]
name = "parquet-s3-blocks-writer"
version = "0.1.0"
description = "A Python app that generates fake data, writes it to a Parquet file in 1MB blocks, and uploads to LocalStack S3"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "parquet_s3_blocks_writer", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
pyarrow = "^18.1.0"
boto3 = "^1.34.0"
faker = "^22.0.0"
python-dotenv = "^1.0.0"
pandas = "^2.2.0"
s3fs = "^2024.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.12.0"

[tool.poetry.scripts]
parquet-s3-writer = "parquet_s3_blocks_writer.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py313']
